+++ js/form.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line no-unused-vars
+import {objectsArray} from './data-generation.js';
